refactor(ai): extract shared logic into AiProvider trait
Extract ~500 LOC of duplicated code between GeminiClient and
OpenRouterClient into a shared AiProvider trait with default
implementations.

Changes:
- Add async-trait dependency for async trait methods
- Create AiProvider trait with default implementations for:
  - send_request() with retry logic
  - analyze_issue() and create_issue() flows
  - Prompt builders (system/user prompts)
- Refactor GeminiClient: 629 LOC -> 123 LOC (80% reduction)
- Refactor OpenRouterClient: 807 LOC -> 180 LOC (78% reduction)
- Update module exports and facade to use trait

This enables easy addition of future providers (Groq, Cerebras)
and prepares for provider chain fallback functionality.

Closes #232
