+++ Cargo.toml
@@ -11,6 +11,7 @@ repository = "https://github.com/clouatre-labs/aptu"
 
 [workspace.dependencies]
 # Core
+async-trait = "0.1"
 clap = { version = "4", features = ["derive"] }
 clap_complete = "4"
 tokio = { version = "1", features = ["full"] }
